keep role-validator tests in test_config.py to stabilize xdist coverage
Moving 13 tests from test_config.py to a new test_role_validator.py
shifted pytest-xdist's loadgroup distribution enough that pre-existing
coverage fragility surfaced: each CI run picked a different single line
in an unrelated file (ner_concepts.py, catalog.py, ...) that nobody on
that particular xdist shard happened to hit. Reverting the file split
takes the worker distribution back to within +2 tests of main, which
keeps the existing coverage stable.
